mittimantra_backend/app/ai_core/llm_gemini_vision_client.py
import os
try:
    import google.generativeai as genai
except ImportError:
    genai = None
from dotenv import load_dotenv

load_dotenv()

class GeminiVisionClient:
    def __init__(self):
        api_key = os.getenv("GEMINI_API_KEY")
        if not api_key:
            self.model = None
        else:
            genai.configure(api_key=api_key)
            # Using 1.5-flash as it's the standard for vision tasks
            self.model = genai.GenerativeModel('gemini-1.5-flash')

    def get_vision_completion(self, prompt, image_data, system_instruction=None):
        if not self.model:
            return "Error: GEMINI_API_KEY not found for vision tasks."
        
        try:
            content = [prompt, image_data]
            if system_instruction:
                prompt_with_sys = f"{system_instruction}\n\n{prompt}"
                content[0] = prompt_with_sys
            
            response = self.model.generate_content(content)
            return response.text
        except Exception as e:
            return f"Gemini Vision Error: {str(e)}"

gemini_vision_client = GeminiVisionClient()
