mittimantra_backend/requirements.txt
# Core FastAPI
fastapi==0.115.5
uvicorn[standard]==0.34.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.20
email-validator>=2.1.0
jinja2

# ML Core
joblib==1.4.2
scikit-learn==1.8.0
numpy==1.26.4

# Deep Learning
tensorflow==2.18.0

# Image Processing
pillow==11.0.0

# Gradient Boosting Models
xgboost==2.1.3
lightgbm==4.5.0
catboost==1.2.7

# Database
sqlalchemy==2.0.35
alembic==1.14.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2

# Environment Variables
python-dotenv==1.0.1
groq
google-generativeai